Version 1.11

+++ sha1.c
@@ -51,12 +51,12 @@ A million repetitions of "a"
 
 /* Hash a single 512-bit block. This is the core of the algorithm. */
 
-void SHA1Transform(unsigned long state[5], unsigned char buffer[64])
+void SHA1Transform(uint32_t state[5], unsigned char buffer[64])
 {
-  unsigned long a, b, c, d, e;
+  uint32_t a, b, c, d, e;
   typedef union {
     unsigned char c[64];
-    unsigned long l[16];
+    uint32_t l[16];
   } CHAR64LONG16;
   CHAR64LONG16* block;
 #ifdef SHA1HANDSOFF
@@ -144,7 +144,7 @@ void SHA1Update(SHA1_CTX* context, unsigned char* data, unsigned int len)
 
 void SHA1Final(unsigned char digest[20], SHA1_CTX* context)
 {
-  unsigned long i, j;
+  uint32_t i, j;
   unsigned char finalcount[8];
 
   for (i = 0; i < 8; i++) {
